Refactor core context (switch to function syntax)
