+++ packages/lib/src/core/Context/CoreContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from 'preact';
-import { CommonPropsTypes } from './CoreProvider';
+import { CommonPropsTypes } from './types';
 import Language from '../../language/Language';
 
 export const CoreContext = createContext({ i18n: new Language(), loadingContext: '', commonProps: {} as CommonPropsTypes });
