+++ packages/lib/src/core/Context/types.ts
@@ -1,0 +1,10 @@
+export interface CommonPropsTypes {
+    isCollatingErrors?: boolean;
+}
+
+export interface CoreProviderProps {
+    children?: any;
+    commonProps?: CommonPropsTypes;
+    i18n: any;
+    loadingContext?: string;
+}
