+++ packages/lib/src/core/Context/useCoreContext.ts
@@ -1,8 +1,6 @@
 import { useContext } from 'preact/hooks';
 import { CoreContext } from './CoreContext';
 
-function useCoreContext() {
-    return useContext(CoreContext);
-}
+const useCoreContext = () => useContext(CoreContext);
 
 export default useCoreContext;
